Resolves Issue #20
Using the native CF method is much less verbose than trying to obtain the byte array from the binary object to pass in to java.nio.file.Files.  We can refactor this later, if you wish, but Issue #20 is a blocker for using cbfs until `create` can accept binary files.